First version interactive menu
